refactor: HomeImg 파일들 경로 이동
